Move reusable functions to util_filters.py, update import
